+++ hexam/src/main/java/com/hexam/dtos/TeacherDTO.java
@@ -10,14 +10,14 @@ import lombok.Data;
 @Data
 @Builder
 public class TeacherDTO {
-    private Integer personId;
+    private Long personId;
     private String profileImage;
     private String fullName;
     private String userName;
     private String emailAddress;
     private Boolean enable;
 
-    public TeacherDTO(Integer personId, String profileImage, String fullName, String userName, String emailAddress, Boolean enable) {
+    public TeacherDTO(Long personId, String profileImage, String fullName, String userName, String emailAddress, Boolean enable) {
         this.personId = personId;
         this.profileImage = profileImage;
         this.fullName = fullName;
